robot-framework/requirements.txt
robotframework >= 7.3.2
robotframework-browser >= 19.6.2
robotframework-tidy >= 4.18.0
robotframework-requests >= 0.9.7
robotframework-jsonlibrary >= 0.5
robotframework-appiumlibrary >= 3.1
python-dotenv >= 1.1.1
